feat: allow saving query as collection
