+++ .vscode/settings.json
@@ -29,5 +29,6 @@
   "typescript.tsdk": "node_modules/typescript/lib",
   "[markdown]": {
     "editor.wordWrap": "on"
-  }
+  },
+  "typescript.preferences.importModuleSpecifier": "non-relative"
 }
